# Added 4 sample techs.
Also modified the entity script to allow for gathering fractional resources, and modified the GUI to round resources to the nearest integer. This is necessary since techs will likely push resource gathering amounts to fractional numbers.

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@4161 3db68df2-c116-0410-a063-a993310a9797

+++ binaries/data/mods/official/scripts/entity_functions.js
@@ -689,9 +689,9 @@ function performGather( evt )
 	}
 
 	if( g.resource[s.type][s.subtype])
-		gather_amt = parseInt( g.resource[s.type][s.subtype] );
+		gather_amt = parseFloat( g.resource[s.type][s.subtype] );
 	else
-		gather_amt = parseInt( g.resource[s.type] );
+		gather_amt = parseFloat( g.resource[s.type] );
 
 	if( s.max > 0 )
 	{
